Solve 'Proverb' on the Rust Exercism track
Iterated over the list vector with .iter().enumerate() to grab hold of each element's index. Firstly the index allowed me to interpolate both the current line and the next line (list[i+1]) of the list with the format! macro. Secondly the index allowed me to do a check to ensure that this would only be done for lines that were not the final line with if list.len() > i+1. In that case I formatted the final line. For each line I used push_str to add it to a string, then returned the final string after the for loop."