Completed resistor color duo in Typescript
Used a private readonly string array to store the colors input. The
constructor contains an if statement to throw an error if less than
two colors are provided. The value function slices the array to its
first two arguments, maps each to its index in the private static
OPTIONS variable, joins them, then casts to number, returning the
value.