Complete space age exercise in Javascript
The ORBITALS constant is assigned to an associative array of planets
and orbital periods relative to Earth years. The
SECONDS_IN_YEARS constant represents the number of seconds in years.
The age function takes a planet and age provided by the user. The
planet is sanitized to lowercase, then handled for errors. If the
user input is "pluto" an error is thrown reminding the user that
Pluto is not a planet. If the user input is anything else not in
the ORBITALS associative array, then a generic error is thrown.

Otherwise, the age is divided by the relative orbital period, which
represents the age of someone on a particular planet in seconds.
This value is then divided by the number of seconds in a year to
attain the age in years rather than seconds. This is then cast to
a two-decimal string with toFixed, and cast back to a floating
point value with Number(). This number is then returned from the
function.