Complete Bank Account Javascript exercise
The bank account class came with the following unimplemented
methods: open, close, deposit, withdraw and a balance getter.

I gave the class an instance property opened, a boolean initially
set to false and used in  method conditionals, ensuring valid bank
operations. The close, deposit, withdraw and balance methods
will throw an error if opened property is set to false. The open
method will open a bank account, and vice versa when the
opened property is set to true; the open method throws an error
and the close method closes the bank account.

The open method not only opens an account but also initializes
an instance property of accountBalance at 0.

Deposit and withdraw ensures the argument passed is >= 0, which
also type checks for a number argument, since the only values
that return true are numbers greater than zero, or arrays with
a single element that is a number greater than zero.

Withdraw ensures the quantity is not greater than the account
balance.

The balance getter returns the accountBalance property that was
initialized by the open method. I found that naming both the
getter "balance" and the account balance property "balance" does
not work, so I called the getter "balance" and the instance
property "accountBalance", which the getter returns when called.