Complete the Matrix Javascript exercise
The Matrix class came with two unimplemented getters, rows and
columns. The goal of the exercise was to implement these two
methods, such that a string could be parsed into a 2D array of
either rows or columns.

To get the rows, we first split the string by new line characters,
so that the string is broken up into an array of strings. We then
split each of these substrings by spaces, so that these substrings
each become a subarray of individual characters. Finally, we cast
each of these individual characters from string type to number
type, giving us a 2D array of digits in the original order of the
string, representing rows.

To get the columns, we follow the above process, but then rebuild
our 2D array using reduce. For each subarray, i.e., each row,
we push the element to a corresponding new subarray in a new 2D
array. For example, if a row were comprised of [1,2,3], these
would be pushed so as to create [[1],[2],[3]], which would be the
beginnings of three column arrays.