Complete Pangram exercise in Javascript
A pangram is a string containing all characters in the alphabet at
least once. To test a string for pangram validity, I build a set of
its unique alphabetical characters, and if the set's size is the
size of the alphabet, then the string is a pangram.

I start by spreading the string into an array which allows me to
call reduce on the characters. I use regex to test if the character
is a letter, and if so, add it to the set. Iff the set's size is
26, the string contained at least one of each letter in the
alphabet, and therefore we return true to represent that the string
is a pangram.