Completed Reverse String in Javascript
Used a recursive function to reverse a string. The base case
returns the string if it is of length 1 or 0. If the string is of
length greater than 1, return the concatenation of its final letter
and the function called on every other string.

This results in a call stack of repeatedly prepending the final
letter to the remaining letters until the end of the string is
reached, at which point we bubble up the call stack, building the
string in reverse order.